Fixed arena instances were instantly destroyed if some cancelled their entry before anyone else could enter.
- Fixed players could teleport into a destroyed instance (if enter was pressed slightly after the other player cancelled)
- Fixed not responsive enter button due to the instance being destroyed

+++ game-server/src/com/aionemu/gameserver/services/AutoGroupService.java
@@ -330,7 +330,7 @@ public class AutoGroupService {
 
 	public boolean destroyIfPossible(AutoInstance autoInstance) {
 		WorldMapInstance instance = autoInstance.getInstance();
-		if (instance.getPlayersInside().stream().noneMatch(Player::isOnline)) {
+		if (autoInstance.getRegisteredAGPlayers().isEmpty() && instance.getPlayersInside().stream().noneMatch(Player::isOnline)) {
 			autoInstances.remove(instance);
 			InstanceService.destroyInstance(instance);
 			return true;
